Changes: Operations folder: • Created subfolders Chat, Damage, Flag • Moved SetFlagToLogLine class file from Operations folders to Flag folder Class UpdateConsoleLogFile • fixed a typo in class name • Added reference to subfolder Flag
